BE:알라딘

+++ backend-java/src/main/java/com/ssafy/db/repository/BookSearchRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.db.repository;
 import com.ssafy.db.entity.BookDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookSearchRepository extends JpaRepository<BookDetail, Long> {
+public interface BookSearchRepository extends JpaRepository<BookDetail, Integer> {
 }
